+++ generator.ts
@@ -7,16 +7,25 @@ import * as os from 'os';
 import * as mustache from 'mustache'
 import * as drillInRpt from './reports/drill-in'
 import * as cp from 'child_process';
+import {ProjectCrawler} from './projectCrawler';
+import {CrawlingConfig, CrawlingTarget} from './interfaces';
 
 let sanitize = require('sanitize-filename');
 let clone = require('clone');
 
-import { GeneratorConfiguration, IssueCard, IssueCardEvent, ReportSnapshot, ReportConfig, ProjectsData, ProjectData, ProjectReportBuilder, ReportDetails } from './interfaces'
+import { GeneratorConfiguration, IssueCard, IssueCardEvent, ReportSnapshot, ReportConfig, ProjectData, ProjectReportBuilder, ReportDetails } from './interfaces'
 
 export async function generate(token: string, configYaml: string): Promise<ReportSnapshot> {
     console.log("Generating reports");
 
-    let configPath = path.join(process.env["GITHUB_WORKSPACE"], configYaml);
+    const workspacePath = process.env["GITHUB_WORKSPACE"];
+    if (!workspacePath) {
+        throw new Error("GITHUB_WORKSPACE not defined");
+    }
+
+    let configPath = path.join(workspacePath, configYaml);
+    let cachePath = path.join(workspacePath, "_reports", ".data");
+
     let config = <GeneratorConfiguration>yaml.load(fs.readFileSync(configPath, 'utf-8'))
 
     let snapshot = <ReportSnapshot>{};
@@ -30,19 +39,51 @@ export async function generate(token: string, configYaml: string): Promise<Repor
     let minute = d.getUTCMinutes().toString().padStart(2, "0");
     let dt: string = `${year}-${month}-${day}_${hour}-${minute}`;
     snapshot.datetimeString = dt;  
-    
-    const workspacePath = process.env["GITHUB_WORKSPACE"];
-    if (!workspacePath) {
-        throw new Error("GITHUB_WORKSPACE not defined");
-    }    
+        
     snapshot.rootPath = path.join(workspacePath, snapshot.config.output);
 
     // apply defaults
     snapshot.config.output = snapshot.config.output || "_reports";
+    await writeSnapshot(snapshot);
 
     // load up the projects, their columns and all the issue cards + events.
-    let projectsData: ProjectsData = await loadProjectsData(token, config, snapshot);
-    console.log("loaded.");
+    let github = new GitHubClient(token, cachePath);
+    let projectData = <ProjectData>{};
+    let crawlCfg: CrawlingConfig;
+    if (typeof(config.targets) === 'string') {
+        throw new Error('crawl config external files not supported yet');
+    }
+    else {
+        crawlCfg = <CrawlingConfig>config.targets;
+    }
+
+    console.log("crawl cfg");
+    console.log(JSON.stringify(crawlCfg, null, 2));
+
+    let crawled = 0;
+    for (let targetName in crawlCfg) {
+        console.log(`target: ${targetName}`);
+        let target:CrawlingTarget = crawlCfg[targetName];
+        if (target.type === 'project') {
+            console.log(`crawling project ${target.htmlUrl}`);
+            let projectCrawler = new ProjectCrawler(github);
+            await projectCrawler.crawl(target, projectData);
+            ++crawled;
+        }
+        else if (target.type === 'repo') {        
+            throw new Error('crawling repos not supported yet');
+            //++crawled;
+        }
+        else {
+            throw new Error(`Unsupported target config: ${target.type}`);
+        }
+    }
+
+    if (crawled == 0) {
+        throw new Error("No targets were crawled for data.  Please specify targets in the config file.");
+    }
+
+    console.log("data crawled.");
 
     // update report config details
     for (const report of config.reports) {
@@ -61,92 +102,86 @@ export async function generate(token: string, configYaml: string): Promise<Repor
         });
     }
 
-    await writeSnapshot(snapshot);
 
-    // hand that full data set to each report to render
-    for (const proj in projectsData) {
-        const projectData = projectsData[proj];
+    for (const report of config.reports) {
+        let output = "";
 
-        for (const report of config.reports) {
-            let output = "";
+        output += getReportHeading(report);
+        console.log();
+        console.log(`Generating ${report.name} ...`);
+        await createReportPath(report);
 
-            output += getReportHeading(report);
-            console.log();
-            console.log(`Generating ${report.name} for ${proj} ...`);
-            await createReportPath(report);
+        for (const reportSection of report.sections) {
+            output += `&nbsp;  ${os.EOL}`;
 
-            for (const reportSection of report.sections) {
-                output += `&nbsp;  ${os.EOL}`;
+            let reportModule = `${reportSection.name}`;
 
-                let reportModule = `${reportSection.name}`;
+            // if it's a relative path, find in the workflow repo relative path.
+            // this allows for consume of action to create their own report sections
+            // else look for built-ins
+            console.log(`Report module ${reportModule}`);
+            let reportModulePath;
 
-                // if it's a relative path, find in the workflow repo relative path.
-                // this allows for consume of action to create their own report sections
-                // else look for built-ins
-                console.log(`Report module ${reportModule}`);
-                let reportModulePath;
+            if (reportModule.startsWith("./")) {
+                reportModulePath = path.join(process.env["GITHUB_WORKSPACE"], `${reportModule}`);
+            }
+            else {
+                reportModulePath = path.join(__dirname, `./reports/${reportSection.name}`);
+            }
 
-                if (reportModule.startsWith("./")) {
-                    reportModulePath = path.join(process.env["GITHUB_WORKSPACE"], `${reportModule}`);
-                }
-                else {
-                    reportModulePath = path.join(__dirname, `./reports/${reportSection.name}`);
-                }
+            console.log(`Loading: ${reportModulePath}`);
 
-                console.log(`Loading: ${reportModulePath}`);
+            if (!fs.existsSync(reportModulePath)) {
+                throw new Error(`Report not found: ${reportSection.name}`);
+            }
 
-                if (!fs.existsSync(reportModulePath)) {
-                    throw new Error(`Report not found: ${reportSection.name}`);
-                }
+            let reportGenerator = require(reportModulePath) as ProjectReportBuilder;
 
-                let reportGenerator = require(reportModulePath) as ProjectReportBuilder;
+            // overlay user settings over default settings 
+            let config = reportGenerator.getDefaultConfiguration();
+            for (let setting in reportSection.config || {}) {
+                config[setting] = reportSection.config[setting];
+            }
 
-                // overlay user settings over default settings 
-                let config = reportGenerator.getDefaultConfiguration();
-                for (let setting in reportSection.config || {}) {
-                    config[setting] = reportSection.config[setting];
-                }
+            console.log("Processing data ...")
 
-                console.log("Processing data ...")
+            let drillIns = [];
+            let drillInCb = (identifier: string, title: string, cards: IssueCard[]) => {
+                drillIns.push({
+                    identifier: identifier,
+                    title: title,
+                    cards: cards
+                })
+            }
 
-                let drillIns = [];
-                let drillInCb = (identifier: string, title: string, cards: IssueCard[]) => {
-                    drillIns.push({
-                        identifier: identifier,
-                        title: title,
-                        cards: cards
-                    })
-                }
+            let processed = reportGenerator.process(config, clone(projectData), drillInCb);
+            await writeSectionData(report, reportModule, config, processed);
 
-                let processed = reportGenerator.process(config, clone(projectData), drillInCb);
-                await writeSectionData(report, reportModule, config, processed);
+            if (report.kind === 'markdown') {
+                output += reportGenerator.renderMarkdown(projectData, processed);
+                // output += `&nbsp;${os.EOL}`;
+            }
+            else {
+                throw new Error(`Report kind ${report.kind} not supported`);
+            }
 
+            for (let drillIn of drillIns) {
+                let drillInReport: string;
                 if (report.kind === 'markdown') {
-                    output += reportGenerator.renderMarkdown(projectData, processed);
-                    // output += `&nbsp;${os.EOL}`;
+                    drillInReport = drillInRpt.renderMarkdown(drillIn.title, clone(drillIn.cards));
                 }
                 else {
                     throw new Error(`Report kind ${report.kind} not supported`);
                 }
 
-                for (let drillIn of drillIns) {
-                    let drillInReport: string;
-                    if (report.kind === 'markdown') {
-                        drillInReport = drillInRpt.renderMarkdown(drillIn.title, clone(drillIn.cards));
-                    }
-                    else {
-                        throw new Error(`Report kind ${report.kind} not supported`);
-                    }
-
-                    await writeDrillIn(report, drillIn.identifier, drillIn.cards, drillInReport);
-                }
+                await writeDrillIn(report, drillIn.identifier, drillIn.cards, drillInReport);
             }
-            console.log("Writing report");
-            writeReport(report, projectData, output);
-            console.log("Done.");
         }
-        console.log();
+        console.log("Writing report");
+        writeReport(report, projectData, output);
+        console.log("Done.");
     }
+    console.log();
 
     return snapshot;
 }
@@ -215,102 +250,4 @@ async function writeReport(report: ReportConfig, projectData: ProjectData, conte
     fs.writeFileSync(path.join(report.details.dataPath, "_project.json"), JSON.stringify(projectData, null, 2));
 }
 
-async function loadProjectsData(token: string, config: GeneratorConfiguration, snapshot: ReportSnapshot): Promise<ProjectsData> {
-    console.log("Querying project data ...")
 
-    let cachePath = path.join(snapshot.rootPath, ".data");
-    let github = new GitHubClient(token, cachePath);
-
-    let projMap = <ProjectsData>{};
-    for (const projHtmlUrl of config.projects) {
-        let proj = await github.getProject(projHtmlUrl);
-        if (!proj) {
-            throw new Error(`Project not found: ${projHtmlUrl}`);
-        }
-
-        projMap[projHtmlUrl] = proj;
-    }
-
-    
-    //console.log(JSON.stringify(projMap, null, 2));
-
-    for (const projectUrl of config.projects) {
-        let project: ProjectData = projMap[projectUrl];
-
-        project.columns = {}
-        let cols = await github.getColumnsForProject(project);
-        cols.forEach((col) => {
-            projMap[projectUrl].columns[col.name] = col.id;
-        })
-
-        
-        let mappedColumns = [];
-        for (const key in config.columnMap) {
-            let colNames = config.columnMap[key];
-            if (!colNames || !Array.isArray) {
-                throw new Error(`Invalid config. column map for ${key} is not an array`);
-            }
-
-            mappedColumns = mappedColumns.concat(colNames);
-        }
-
-        let seenUnmappedColumns: string[] = [];
-        project.stages = {}
-        for (const key in config.columnMap) {
-            project.stages[key] = [];
-
-            console.log(`Processing stage ${key}`);
-            let colNames = config.columnMap[key];
-
-            for (const colName of colNames) {
-                let colId = projMap[projectUrl].columns[colName];
-
-                // it's possible the column name is a previous column name
-                if (!colId) {
-                    continue;
-                }
-
-                let cards = await github.getCardsForColumns(colId, colName);
-
-                for (const card of cards) {
-                    // called as each event is processed 
-                    // creating a list of mentioned columns existing cards in the board in events that aren't mapped in the config
-                    // this will help diagnose a potential config issue much faster
-                    let eventCallback = (event: IssueCardEvent):void => {
-                        let mentioned;
-                        if (event.project_card && event.project_card.column_name) {
-                            mentioned = event.project_card.column_name;
-                        }
-
-                        if (event.project_card && event.project_card.column_name) {
-                            mentioned = event.project_card.previous_column_name;
-                        }                        
-
-                        if (mentioned && mappedColumns.indexOf(mentioned) === -1 && seenUnmappedColumns.indexOf(mentioned) === -1) {
-                            seenUnmappedColumns.push(mentioned);
-                        }
-                    }
-
-                    // cached since real column could be mapped to two different mapped columns
-                    // read and build the event list once
-                    let issueCard = await github.getIssueForCard(card, project.id);
-                    if (issueCard) {
-                        util.processCard(issueCard, project.id, config, eventCallback);
-                        project.stages[key].push(issueCard);
-                    }
-                }
-            }
-        }
-        console.log("Done processing.")
-        console.log();
-        if (seenUnmappedColumns.length > 0) {
-            console.log();
-            console.log(`WARNING: there are unmapped columns mentioned in existing cards on the project board`);
-            seenUnmappedColumns = seenUnmappedColumns.map(col => `"${col}"`);
-            console.log(`WARNING: Columns are ${seenUnmappedColumns.join(" ")}`);
-            console.log();
-        }
-    }
-
-    return projMap;
-}
