+++ projectCrawler.ts
@@ -1,0 +1,240 @@
+import {GitHubClient} from './github'
+import {CrawlingTarget, ProjectData, IssueCard, IssueCardEvent} from './interfaces';
+
+let stageLevel = {
+    "None": 0,
+    "Proposed": 1,
+    "Accepted": 2,
+    "In-Progress": 3,
+    "Blocked": 4,
+    "Done": 5
+}
+
+
+export class ProjectCrawler {
+    github: GitHubClient;
+
+    // cache the resolution of stage names for a column
+    // a columns by stage names are the default and resolve immediately
+    columnMap = {
+        "proposed": "Proposed",
+        "accepted": "Accepted",
+        "in-progress": "In-Progress",
+        "done": "Done",
+        "blocked": "Blocked"
+    }
+
+    // keep in order indexed by level above
+    stageAtNames = [
+        'none',
+        'project_proposed_at',
+        'project_accepted_at',
+        'project_in_progress_at',
+        'project_blocked_at',
+        'project_done_at'
+    ]
+
+    constructor(client: GitHubClient) {
+        this.github = client;
+    }
+
+    public async crawl(target: CrawlingTarget, projectData: ProjectData): Promise<void> {
+        console.log(`Crawling project ${target.htmlUrl} ...`);
+
+        let columns: { [key: string]: number } = {};
+
+        let proj = await this.github.getProject(target.htmlUrl);
+
+        let cols = await this.github.getColumnsForProject(proj);
+        cols.forEach((col) => {
+            columns[col.name] = col.id;
+        })
+
+        let mappedColumns = [];
+        for (const key in target.columnMap) {
+            let colNames = target.columnMap[key];
+            if (!colNames || !Array.isArray) {
+                throw new Error(`Invalid config. column map for ${key} is not an array`);
+            }
+
+            mappedColumns = mappedColumns.concat(colNames);
+        }
+
+        let seenUnmappedColumns: string[] = [];
+        projectData.stages = {}
+        for (const key in target.columnMap) {
+            projectData.stages[key] = [];
+
+            console.log(`Processing stage ${key}`);
+            let colNames = target.columnMap[key];
+
+            for (const colName of colNames) {
+                let colId = columns[colName];
+
+                // it's possible the column name is a previous column name
+                if (!colId) {
+                    continue;
+                }
+
+                let cards = await this.github.getCardsForColumns(colId, colName);
+
+                for (const card of cards) {
+                    // called as each event is processed 
+                    // creating a list of mentioned columns existing cards in the board in events that aren't mapped in the config
+                    // this will help diagnose a potential config issue much faster
+                    let eventCallback = (event: IssueCardEvent):void => {
+                        let mentioned;
+                        if (event.project_card && event.project_card.column_name) {
+                            mentioned = event.project_card.column_name;
+                        }
+
+                        if (event.project_card && event.project_card.column_name) {
+                            mentioned = event.project_card.previous_column_name;
+                        }                        
+
+                        if (mentioned && mappedColumns.indexOf(mentioned) === -1 && seenUnmappedColumns.indexOf(mentioned) === -1) {
+                            seenUnmappedColumns.push(mentioned);
+                        }
+                    }
+
+                    // cached since real column could be mapped to two different mapped columns
+                    // read and build the event list once
+                    let issueCard = await this.github.getIssueForCard(card, projectData.id);
+                    if (issueCard) {
+                        this.processCard(issueCard, projectData.id, target, eventCallback);
+                        projectData.stages[key].push(issueCard);
+                    }
+                }
+            }
+        }
+
+        console.log("Done processing.")
+        console.log();
+        if (seenUnmappedColumns.length > 0) {
+            console.log();
+            console.log(`WARNING: there are unmapped columns mentioned in existing cards on the project board`);
+            seenUnmappedColumns = seenUnmappedColumns.map(col => `"${col}"`);
+            console.log(`WARNING: Columns are ${seenUnmappedColumns.join(" ")}`);
+            console.log();
+        }
+    }
+
+    // process a card in context of the project it's being added to
+    // filter column events to the project being processed only since. this makes it easier on the report author
+    // add stage name to column move events so report authors don't have to repeatedly to that
+    public processCard(card: IssueCard, projectId: number, target: CrawlingTarget, eventCallback: (event: IssueCardEvent) => void): void {
+        let filteredEvents = [];
+
+        // card events should be in order chronologically
+        let currentStage: string;
+        let doneTime: Date;
+        let blockedTime: Date;
+        let addedTime: Date;
+
+        if (card.events) {
+            for (let event of card.events) {
+                // since we're adding this card to a projects / stage, let's filter out
+                // events for other project ids since an issue can be part of multiple boards
+                
+                if (event.project_card && event.project_card.project_id !== projectId) {
+                    continue;
+                }
+
+                eventCallback(event);
+                
+                let eventDateTime: Date;
+                if (event.created_at) {
+                    eventDateTime = event.created_at;
+                }
+
+                // TODO: should I clear all the stage_at datetimes if I see
+                //       removed_from_project event?
+
+                let toStage: string;
+                let toLevel: number;
+                let fromStage: string;
+                let fromLevel: number = 0;
+
+                if (event.project_card && event.project_card.column_name) {
+                    if (!addedTime) {
+                        addedTime = eventDateTime;
+                    }
+
+                    toStage = event.project_card.stage_name = this.getStageFromColumn(event.project_card.column_name, target);
+                    toLevel = stageLevel[toStage];
+                    currentStage = toStage;
+                }
+        
+                if (event.project_card && event.project_card.previous_column_name) {
+                    fromStage = event.project_card.previous_stage_name = this.getStageFromColumn(event.project_card.previous_column_name, target);
+                    fromLevel = stageLevel[fromStage];
+                }
+
+                // last occurence of moving to these columns from a lesser or no column
+                // example. if moved to accepted from proposed (or less), 
+                //      then in-progress (greater) and then back to accepted, first wins            
+                if (toStage === 'Proposed' || toStage === 'Accepted' || toStage === 'In-Progress') {
+                    if (toLevel > fromLevel) {
+                        card[this.stageAtNames[toLevel]] = eventDateTime;
+                    } 
+                }
+
+                if (toStage === 'Done') {
+                    doneTime = eventDateTime;
+                }
+
+                if (toStage === 'Blocked') {
+                    blockedTime = eventDateTime;
+                }
+
+                filteredEvents.push(event);
+            }
+            card.events = filteredEvents;
+
+            // done_at and blocked_at is only set if it's currently at that stage
+            if (currentStage === 'Done') {
+                card.project_done_at = doneTime;
+            }
+
+            if (currentStage === 'Blocked') {
+                card.project_blocked_at = blockedTime
+            }
+
+            if (addedTime) {
+                card.project_added_at = addedTime;
+            }
+
+            card.project_stage = currentStage;
+        }
+    } 
+    
+    private getStageFromColumn(column: string, target: CrawlingTarget): string {
+        column = column.toLowerCase();
+        if (this.columnMap[column]) {
+            return this.columnMap[column];
+        }
+
+        let resolvedStage = null;
+        for (let stageName in target.columnMap) {
+            // case insensitve match
+            for (let mappedColumn of target.columnMap[stageName]) {
+                let lowerColumn = mappedColumn.toLowerCase();
+                if (lowerColumn === column.toLowerCase()) {
+                    resolvedStage = stageName;
+                    break;
+                }
+            }
+
+            if (resolvedStage) {
+                break;
+            }
+        }
+
+        // cache the n^2 reverse case insensitive lookup.  it will never change for this run
+        if (resolvedStage) {
+            this.columnMap[column] = resolvedStage;
+        }
+
+        return resolvedStage;
+    }    
+}
